requirements.txt
# ============================================================================
# Onelab Payments Reconciliation — Python dependencies
# Tested with Python 3.11+
# ============================================================================

# ---------------------------------------------------------------------------
# [CORE] — Data engine and web framework
# ---------------------------------------------------------------------------
streamlit==1.39.0          # multi-page web UI framework — the dashboard itself
pandas==2.2.3              # vectorised reconciliation joins & aggregations
numpy==2.1.3               # numeric backend for pandas / random sampling

# ---------------------------------------------------------------------------
# [VISUALIZATION] — Interactive charts
# ---------------------------------------------------------------------------
plotly==5.24.1             # interactive charts (gauge, donut, waterfall, heatmap)

# ---------------------------------------------------------------------------
# [DATA GENERATION] — Synthetic financial data
# ---------------------------------------------------------------------------
faker==30.8.2              # realistic-looking synthetic data primitives

# ---------------------------------------------------------------------------
# [UTILITIES] — Configuration & I/O
# ---------------------------------------------------------------------------
python-dotenv==1.0.1       # load .env config (seed, volume, recon period)

# ---------------------------------------------------------------------------
# [TESTING] — Test runner
# ---------------------------------------------------------------------------
pytest==8.3.3              # unit + integration tests for engine & gap detection
